refactor(settings) separate out components
